Assignment for Advanced Programming: Concurrency
Java assignment for advanced programming concurrency. Java model of
scenario where "Martin" a character can not leave his house and make use
of the road if there are cars already using the road. Locking the road
and keeping track of how many objects are on the road at any one time to
prevent concurrent issues.